requirements.txt
fastapi
requests
json5
langchain==0.0.154
openai==0.28
python-dotenv
pdfminer.six
faiss-cpu
langchain-community==0.0.38
tiktoken
pydantic==1.10.7
youtube-transcript-api